pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsfeed"
version = "0.2.0"
description = "Simple command line news aggregator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "platformdirs>=4,<5",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "python-dateutil==2.9.0",
    "blessed==1.20.0",
    "nltk==3.9.1",
]

[project.scripts]
newsfeed = "newsfeed:main"

[tool.setuptools]
py-modules = ["newsfeed", "newsfeed_app_config"]

[tool.setuptools.packages.find]
include = ["app*", "newsfeed_config*"]

[tool.setuptools.package-data]
newsfeed_config = ["config.default.json"]
